feat: refactor Room service and event handling for improved stage management
