+++ src/main/java/com/icebreaker/be/application/room/event/RoomStageEventListener.java
@@ -2,8 +2,8 @@ package com.icebreaker.be.application.room.event;
 
 
 import com.icebreaker.be.application.room.RoomStageService;
-import com.icebreaker.be.application.room.notify.RoomStageWebSocketNotifier;
 import com.icebreaker.be.global.annotation.AsyncTransactionalEventListener;
+import com.icebreaker.be.infra.messaging.room.RoomStageWebSocketNotifier;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -12,7 +12,7 @@ import org.springframework.transaction.event.TransactionPhase;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class RoomEventListener {
+public class RoomStageEventListener {
 
     private final RoomStageService roomStageService;
     private final RoomStageWebSocketNotifier notifier;
